% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convertFactor.R
\name{convertFactor}
\alias{convertFactor}
\title{Factor converter}
\usage{
convertFactor(x, name)
}
\arguments{
\item{x}{Vector of string or factors}

\item{name}{Name of vector supplied}
}
\value{
\code{convertFactor} returns a data frame with indicators for each of the factored groups.
}
\description{
\code{convertFactor} converts factored/character vectors into a data frame of indicators, with the reference group being the largest.
}

