% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretizing longitudinal data}
\usage{
discretize(dataset, date, startdate, stopdate, outcome, range)
}
\arguments{
\item{dataset}{Longitudinal data set to be discretized in \code{data.frame} format.}

\item{date}{Covariate name for date variable.}

\item{startdate}{Covariate name for the date started from for discretization.}

\item{stopdate}{Covariate name for the date ending from for discretization.}

\item{outcome}{Covariate name for whether an event or censoring occurs. Must be one of the following: death, ltfu, eofu, transfer}

\item{range}{Length in days of each interval}
}
\value{
\code{discretize} returns a data.frame with the discretized data and variables corresponding to the visits conducted.
}
\description{
\code{discretize} breaks longitudinal data into equivalent time intervals using the last observed visit leading up to each new interval.
}
\details{
In addition to breaking longitudinal data into equally spaced time intervals, the code will return a data frame with the number of visits observed in the interval leading up to the current interval and within the current interval, the most recent visit observed leading up to the current interval, and the next visit expected.
}

