% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LTFUvisit.R
\name{Lastvisit}
\alias{Lastvisit}
\title{Last observation before censoring}
\usage{
Lastvisit(x, limit)
}
\arguments{
\item{x}{An ordered vector of observed visits.}

\item{limit}{The maximum number of days that can pass before the subject is censored.}
}
\value{
\code{Lastvisit} returns the last observed visit prior to being censored.
}
\description{
\code{Lastvisit} gets the last observation before censoring at the user specified day limit.
}
\details{
This code assumes that the date vector is already in order and will return the last visit date before the specified number of days have passed.
}

