% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/limitedImpute.R
\name{limitedImpute}
\alias{limitedImpute}
\title{Limited last observation carried forward imputation}
\usage{
limitedImpute(var, date, limit)
}
\arguments{
\item{var}{Variable to be imputed forward in time.}

\item{date}{Date variable corresponding to when \code{var} was measured.}

\item{limit}{Number of days the forward imputation is limited to.}
}
\value{
The function returns a vector of length \code{var}, with the limited imputation carried out.
}
\description{
\code{limitedImpute} will carry forward the last observation up to a specified threshold of days.
}
\details{
This function is meant for indicators with missing values known to only occur for a limited amount of time. Examples include pregnancy (45-weeks) and tuberculosis treatment (8 months). Once the specified threshold has passed, the indicator reverts back to 0. If the values are non-missing, the value gets used regardless of whether it surpasses the threshold.
}

