% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nadir.R
\name{nadir}
\alias{nadir}
\title{Running nadir value}
\usage{
nadir(x)
}
\arguments{
\item{x}{Vector to take running nadir over.}
}
\value{
\code{nadir} returns an vector of the nadir values.
}
\description{
\code{nadir} Takes the running lowest value in vector of values pre-ordered by time.
}
\details{
This function assumes that the values in the supplied vector have already been ordered chronologically.
}

