% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rawDataSummary.R
\name{rawDataSummary}
\alias{rawDataSummary}
\title{Summary of raw data}
\usage{
rawDataSummary(file)
}
\arguments{
\item{file}{Location of the excel file to export summaries to.}
}
\value{
\code{rawDataSummary} returns nothing.
}
\description{
\code{rawDataSummary} Summarizes raw data and exports data to an Excel results file.
}
\details{
File needs to be created beforehand with a sheet with the name "Raw data".
}

