% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zenith.R
\name{zenith}
\alias{zenith}
\title{Running zenith value}
\usage{
zenith(x)
}
\arguments{
\item{x}{Vector to take running zenith over.}
}
\value{
\code{zenith} returns an vector of the zenith values.
}
\description{
\code{zenith} Takes the running highest value in vector of values pre-ordered by time.
}
\details{
This function assumes that the values in the supplied vector have already been ordered chronologically.
}

